backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==1.2.0
pinecone-client==3.2.2
sentence-transformers==3.0.1
